+++ lib/versao.ts
@@ -1,3 +1,3 @@
-export const VERSAO = '0.3.3';
+export const VERSAO = '0.3.4';
 
 export const VERSAO_NOME = 'Sessão que dura';